refactor: move QuivrLogo to lib
