+++ frontend/lib/assets/QuivrLogo.tsx
